requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.20.0
PyYAML>=6.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0
pytorch-lightning>=2.0.0

# Data handling and preprocessing
pandas>=1.5.0
h5py>=3.8.0
scipy>=1.10.0

# Metrics and evaluation
scikit-learn>=1.2.0
mir_eval>=0.7

# Development tools
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.1
mypy>=1.3.0

# WavTokenizer specific dependencies
einops>=0.6.1
transformers>=4.30.0
huggingface_hub>=0.16.4
omegaconf>=2.3.0
hydra-core>=1.3.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0  # Optional: for experiment tracking

# Optional: for notebook development
jupyter>=1.0.0
ipython>=8.12.0
